[FIX] stock: protect quant adjust diff at creation
Issue:

When creating new quants, a non correct `inventory_diff_quantity` might
be set outside of inventory mode. This is usually unseen as the field
is invisible in the inventory quant view unless manually set but then
the set value erase the error.

Steps to reproduce:

-Enable Multi-steps route
- Create a receipt for 5 unit of a storable product that is not in stock
- Mark as Todo  and validate the receipt Validate
- Go to Inventory Adjustments and group by product
> on the summary line for quants of the product, the total diff is -5

Fix:

We force the value at creation if the quant was not created from the
inventory mode.

opw-3944310