[FIX] website: make `s_numbers_charts` graph resize properly
This commit aims to make the graph in `s_numbers_charts` resize
simultaneously with it's col parent.

It looks like the problem came from the fact that the `s_chart` div was
composed of the chart alone. Adding some other type of content makes the
chart resize properly.

This commit also fixes an implementation issue that was not reproducible
with the editor (`<p class="display-1">`).

task-4207620