[IMP] web_editor, website: image as social media icons
Before this commit, it was possible to replace social media icon with
any image. However, the image was not displayed as an icon and did not
behave like the others (for example it would affect the display because
the size of the image was bigger). The options in the editor were the
ones for an image.

This commit allows the use of images as icons with the correct behavior.

task-3609071