[FIX] point_of_sale: prevent 0 amount payment lines
Steps to reproduce :
Install the point_of_sale module.
Order something and go to payment page.
Apply multiple payment lines even with 0 amount.

Issue :
The payment lines with 0 amount is also created on the backend
and  shown on receipt as well.

Cause :
The applied condition was wrong to remove paymentline.

Fix :
Updated the condition to remove 0 amount paymentline.