[FIX] website: fix minor issues within the resource editor
This commit aims to fix minor design issues within the resource editor
that require a small amount of changes for a high gain.

1) The `save` and `close` buttons in the top right corner are missing
a spacing, making them stick together ;

2) When opening the resource editor, a modal is displayed to indicate
that the changes applied here could be lost during an upgrade.

The content of that modal was missing a bit of spacing, and there was an
issue with the layout of the bottom buttons.

This commit takes care of these changes while keeping the same content.

task-4008742