[PERF] stock: group similar inventory adjustment moves
Previously, when making inventory adjustments, a move was created for
each modified line (quant) in the inventory adjustment view.
This is suboptimal for cases where we need to adjust large amounts of
similar quants (for example 1000 different serial numbers of the same
product in the same location).
To speed up this use case and limit the number of stock_move/svl entries
that are created, we now only create one move for each product-wh combo,
with one move line per modified item (quant) in the inventory adjustment
view.

task-4241494