[IMP] sale: no mail on order cancellation
Prior to this commit:
- When a quote or a sales order (SO) was canceled using the cancel button in the
  view, an email template wizard would open.
- There were two purposes behind this:
  1. To prevent users from canceling an SO inadvertently.
  2. To notify a customer when an order is canceled.

After this commit:
The process of canceling an SO has been simplified by removing the email
template but retaining the warning. It is important to prevent users from
canceling orders by mistake, but if a user needs to send a message to the
customer to notify the cancellation, it's probably because they have already
had some discussion with the customer. In such cases, they can simply use the
chatter to mention that the order has been cancelled.

task-4212816