[ADD] base: Validation for Action-View Mapping
Cause: The current flow allows users to create action-to-view
mappings for a specific `view_mode`, even when no `view_id`
is provided and no default view of the specified type exists
for the particular model.

This issue prevents customers from accessing other available
`views` mapped for the action, even though valid views are
present.
https://github.com/odoo/odoo/blob/e991b5813e17cf2c0a70498a819208feed24cb11/odoo/addons/base/models/ir_ui_view.py#L2540

Solution: After this patch, an error will be shown to users
at the time of mapping creation, instead of waiting for them
to encounter the issue when accessing the action later.