[FIX] core: correctly handle deferred constraints
When a module redefines a constraint the ORM fails to add it. It puts
the operation in a deferred queue and tries to apply them after all
modules are loaded. This causes issues with constraints redefinition.
E.g. if module A defines a constraint that's overridden in module B and
some data is added in the DB that violates the constraint in A. Then
each time we reload the DB the ORM would try to re-add the constraint in
A and it will fail each time.

Up to now we just output a warning in the logs when the constraint
cannot be added in the post processing of the constraints. In this patch
we propose to remove any pending operation for a constraint with the
same name, ignoring the definition. This would essentially make an
override properly work. Under the assumption that when a constraint is
being applied, any previous one that failed (and are queued) with the
same name can be ignored since it's being overridden.

Example in current modules: the redefinition of `unique_login`
constraint in `res.users` in `website` module,
https://github.com/odoo/odoo/blob/09fa1db600c31d2d04f0504f0412cbbc85ff7e3d/addons/website/models/res_users.py#L18
overriding the definition in `base`.
https://github.com/odoo/odoo/blob/09fa1db600c31d2d04f0504f0412cbbc85ff7e3d/odoo/addons/base/models/res_users.py#L414

If we create two users with same login but different websites the
original constraint in base cannot be applied and it fails in
`finalize_constraints`
https://github.com/odoo/odoo/blob/09fa1db600c31d2d04f0504f0412cbbc85ff7e3d/odoo/modules/registry.py#L574-L577

```
odoo.schema:577 Table 'res_users': unable to add constraint 'res_users_login_key' as UNIQUE (login)
```