[FIX] website_sale: contact widget safari shipping
Same issue solved by commit 6abbb443c1b64580f6dc246de3c102cbbf6dd6c3
In a different component.
Contact widget was causing a schema.org error on safari for mobile versions
in checkout and in payment confirmation, which caused customers to be weary
of paying on odoo.
Removed the use of contact widget and displayed the address instead

opw-4128571