[IMP] core: improve convert_to_column_update
`convert_to_column_update` can trust the data in cache and doesn't need
`convert_to_write` and `validate`

take more advantage of `validate` for converters to avoid validating trusted or
already validated data

For `Properties` fields, format data when `convert_to_cache` for write instead
of `convert_to_read` for flush, which also improves the data consistency. As a
result, it is required to update the property definition before update the
property value.