[FIX] account_edi_ubl_cii: enable factur-x for custom PDF reports
It is currently not possible to embed the factur-x.xml in custom pdf
templates, although it is supported in 17.4.

In this commit, we add the possibility to to do so, by adding a system
parameter "account.custom_templates_facturx_list" that contains a list
of the custom report names (field `report_name` on `ir.actions.report`)
for which we want to embed a factur-x.xml.

We also re-introduce the extension of
`_render_qweb_pdf_prepare_streams`. In addition, we add a context_key
`from_account_move_send` to avoid adding the Factur-X twice in case we
are in the Send & Print flow (see
`_hook_invoice_document_after_pdf_report_render`).

opw-4160945

X-original-commit: 04a827e553cd6389a366c02d3a556de8c657ee22