[FIX] website: fallback if company email is empty
If the non-mandatory Email field is not set
on the res.company, the website form submission uses
a <False> value for the email address
(request.env.company.email).

To avoid this, fall back on a mandatory email
that is configured on the website_editor
when the form action set is to "Send an E-mail".

Additionally, log the error.
Although the error can easily be fixed in the front-end
by simply setting an email for the company,
we should not rely on a non-mandatory field.

opw-4215851
opw-4148791
opw-4160591