[FIX] sale_project: make sure the modal is closed before doing next step
Before this commit, the `task_create_sol_tour` could fail before the
step did not wait enough the modal be closed before saving the form and
so the form view could not be saved before going in the project app.

This commit adds a step to make sure the modal is closed and does the
step to save the form view when we check everything in the form view to
make sure the form view will be saved before leaving the view.

runbot-77457