[FIX] base: temporarily fix duplicate extracted emails
When we create a partner using only an email, its name is set to the
given email. This leads to a formatted addreee looking like
'"Bob@test.com" <bob@test.com>' which leads IrMailServer to send
2 emails when going through 'extract_rfc2822_addresses'. Indeed
it finds an email in name and consider it as a valid smtp to.

As a temporary workaround until a proper solution is found we already
uniquify the result of 'extract_rfc2822_addresses' when preparing
SMTP to.

Task-3704658: [mail] Avoid sending email on partner name