[FIX] l10n_eu_oss: tax mappings in a hierarchy unique per tax id
Currently creating the tax mappings for a branch company creates the tax
mappings in the first parent company (including the branch) that has
a tax id.
Multiple companies in a hierarchy may share the same tax id.
E.g. let B be a branch of C. Then B would have access to the
mappings created for C. There is no need to also create the tax
mappings for B.

After this commit we only create a tax mapping for a company
if none of the parent companies have the same tax id.
We also ensure that the needed tax mappings in the parent companies are created.

task-4247724