[IMP] point_of_sale: number buffer qty sign change when pressing +/-
Before this commit:
Previously, when adding a product and clicking the "+/-" button, the quantity
would change to "0" instead of becoming negative.

After this commit
After implementing this commit, pressing the +/- button now results in the
quantity value changing to negative instead of changing it to 0.

task- 4281516
Related PR: https://github.com/odoo/enterprise/pull/73653