[FIX] account: do not mark payment paid if no reco required
Currently, a payment without outstanding is set as paid as soon as the
invoice is paid, most likely through the reconciliation widget.
However, the decision was made to not require the reconciliation when
using the enterprise Invoicing[^1].

These two facts are conflicting with batch payments since they can be
used in Invoicing, but a payment cannot be added to a batch payment if
it is already paid.

The fix is to always keep a payment as "in process" if the full
Accounting is not installed. They will still be marked if the user sets
it manually, or through the batch payment when it is validated.

runbot-104545

[^1]: odoo/enterprise@adfe6c2d5feb431043e78c423748ab87fae64084