[IMP] html_editor: add toggle list
This commit adds a new type of lists: toggle lists.
These type of lists enable the users to hide text inside a toggle which frees a
lot of space inside the html field.

These toggles can be used to have a more complete description for a specific
title. Or to hide some details about specific concepts.

On a technical point, this commit adds a new embedding and a plugin.
The embedding is quite a simple one as the element itself is simple.
The plugin on the other hand needed to handle a big part of editor behaviors:
Enter, Shift-tab, Tab, etc.
That is why we added new resources to make it easy to handle some specific
behaviors.

task-2796594