[FIX] project_todo: add notification error if project not set
Before this commit, when the user would like to convert a todo into a
task, a form view dialog will be displayed to let him to set some fields
such tags and project. The problem is if the project is not set, the
conversion cannot really be done.

This commit adds a danger notification saying the project has to be
defined to be able to convert a todo into a task.

task-4190723