[FIX] sale: fix tracking of field on quotation
version: 17.0
Issue:
We are not able to track the changes of the field  of a Sale order,
when it is in the "Quotation" stage.

Cause:
Tracking was only triggered after the order was confirmed or it will move to
'Quotation sent' stage,

fix:
Ensure that field tracking is enabled for Sale Orders even when
they are in the "Quotation" stage.

opw-4260555