[FIX] hr_holidays: fix accrual plan in hours
Steps
- create an accrual plan with a level giving 1 hour every month and a carry
over with a maximum of 4 hours. Set accrued gain time at the start of the period
and carry-over time at start of the year.
- create an allocation using this accrual plan, for a time off type in days,
set the start date on the first day of next year.
- go to Time Off Dashboard
- check the balance on the middle of the month of the following year (e.g.
15 january 2026 if the allocation starts on 1 january 2025).

-> ~14 days of time off available: the max postpone amount is not applied,
the time gained is in days and not in hours and more than 1 day is gained/month
(balance a few days after the start of the allocation should be +1 added_value
but is more than that).

Causes:
- the maximum amount of carry over `postpone_max_days` is applied in days
even if the accrual `added_value` is in hours and the UI shows `postpone_max_days`
as "Up to X hours".
- `get_future_leaves` returns a number of hours if the allocation is in hours, even
if the time off type is in days.
- the number of hours from the accrual at the beginning of the allocation is too
high if the start of the accrual is in the future and gain time added at the start
of the period due to cache not being invalidated.
https://github.com/odoo/odoo/blob/19c7737fc05539079cde8641bab608c605efacc5/addons/hr_holidays/models/hr_leave_allocation.py#L575-L578

opw-4272315