[FIX] core: unsuck upgrade code cli
Moved the logs/progress on stderr so it is possible to silent them using
a shell redirection `2>/dev/null`. Simplified what is written on stdout
so it is scriptable.

Moved `sys.stdout.isatty()` outside of `print_progress` so the function
is evaluated once, and not everytime we print a progress. Should speedup
the scripts a bit on runbot.

Moved the `print_progress` in scripts inside the loop, so it doesn't try
to `print_progress(..., 0)` and crash on a division by 0. The runner now
takes care of logging 0% and 100%, along with the script name.

Changed `-t/--test` for `--dry-run`, for consistency with other posix
commands such as rsync.

Made so it exits 1 when there are dirty files.

Added a `--script` option to run a single script. Useful in development
when we want to test a single script and not run all the scripts of a
given version.

Removed the default of `--from`, we couldn't agree on a sensitive
default value. `--from` (or `--script`) is now required.

Used mature `parse_version` from the tools, which is suitable to parse
odoo versions.

Added a long text that describes what "code upgrade" is. It is used as
the module docstring and the --help description.

Used cross-plateform \r instead of unix specific \033.

Skipped `__pycache__` folders, not strickly necessary as `.pyc` files
would nevertheless be skipped via `AVAILABLE_EXT`.

Replaced some string manipulation on paths by pathlib.

X-original-commit: dc6bf09f857f7c26969c2672f7fbaea027997cfb