[IMP] website: add desktop alignment option for sidebar header
With this commit [1], we lost the functionality to change the sidebar
header position, making it impossible to align the sidebar header to the
right. Previously, we managed the sidebar position and text alignment
using the common option, `Format > Alignment`.

In this commit, we introduce a new option, `Desktop Alignment` for
setting the sidebar header position, while keeping `Format > Alignment`
for text alignment to simplify the user interaction with these options.

[1]: https://github.com/odoo/odoo/commit/f87cee6d589694b56caef84beb90259db6cf96b0

task-4204213