[FIX] payment_mercado_pago: round payment tx amounts for COP, HLN & NIO
Versions
--------
- 17.0+

Steps
-----
1. Use COP, HLN or NIO as currency;
2. create a SO or invoice with a non-integer amount total;
3. create a payment link;
4. pay using Mercado Pago.

Issue
-----
> Error processing payment
> Prices in COP currency must be expressed with integer values

Cause
-----
Our currency defaults follow the ISO 4217 standard for minor units. For
three currencies (COP, HLN & NIO), Mercado Pago only accepts integer
amounts instead of the standard 2 decimals, resulting in a failed
payment transaction.

Solution
--------
When creating a transaction, check conditions on whether rounding is
needed. If so, round the amount, as well as the SO and/or invoice.

For invoices, `account.cash.rounding` records can be used to allow for
different strategies: post difference in a journal, or add it to the
biggest tax line.

For sale orders, default to the biggest tax line strategy.

opw-4191877