[IMP] base: cache frozenset instead of tuple
We don't care about the order of `env.companies` in the context of
`__accessible_branches`.
This improvement avoids caching a value for all the permutations of
`env.companies`, slightly reducing the number of queries and the size of
the cache[^1]

[^1]: It is slightly more expensive for the size if the set of
companies is small, but it becomes a lot less expensive the more
possible permutation are possible.